refactor(Controllers): move GetFiles function from ComparesController.cs to FilesController.cs
