+++ src/Ml.Cli.WebApp/Controllers/FilesController.cs
@@ -1,4 +1,7 @@
-﻿using System.Threading.Tasks;
+﻿using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
 using System.Web;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -21,6 +24,25 @@ namespace Ml.Cli.WebApp.Controllers
             _fileLoader = fileLoader;
             _basePath = basePath;
             _comparesPaths = comparesPaths;
+        }
+        
+        [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(IEnumerable<string>))]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> GetFiles()
+        {
+            var jsonExtension = ".json";
+            if (_comparesPaths.Paths == string.Empty)
+            {
+                return BadRequest("Compare repositories paths is unspecified.");
+            }
+            var paths = _comparesPaths.Paths.Split(Separators.CommaSeparator);
+            var fullyQualifiedPaths = paths.Select(path => Path.IsPathRooted(path) ? path : Path.Combine(_basePath.Path, path));
+            
+            var jsonsList = fullyQualifiedPaths
+                .SelectMany(path => _fileLoader.EnumerateFiles(path))
+                .Where(file => Path.GetExtension(file) == jsonExtension);
+            return Ok(jsonsList);
         }
 
         [HttpGet("{id}")]
